requirements.txt
torch>=1.12
pytorch_lightning==1.8.6
ultralytics>=8.3.12
cython
fastapi
setuptools
SQLAlchemy
python-decouple
asyncpg
realesrgan
aiofiles
uvicorn
numpy
opencv_python
scikit_image
asyncio
gitpython
pycocotools
ujson
pillow
tqdm
matplotlib
scipy
PyYAML>=5.4
seaborn
ipywidgets
gevent
termcolor
scikit-learn
albumentations
python-multipart
torchvision>=0.13.1
alembic

